failover-demo2/mysql-consumer/requirements.txt
kafka-python==2.0.2
mysql-connector-python==8.0.26
prometheus-client==0.16.0
faker==19.3.1
requests==2.31.0